This is refactored to separate the league helper methods and clean up our code.
